+++ CHANGES.md
@@ -4,8 +4,11 @@ Changes
 1.3.5.10 (unreleased)
 ---------------------
 
-- Nothing changed yet.
+- Careplan.activity.type and careplan.activity.description have been declared
+  obsolete. This means we make careplan.activity.type optional now to remain
+  backwards compatible as much as possible.
 
+  see https://github.com/Koppeltaal/Koppeltaal-Documentation/issues/36
 
 1.3.5.9 (2021-01-21)
 --------------------